move map to base class
